Move utf-8 validating helpers to new mod
